+++ packages/core_engine/src/datalex_core/dbt/__init__.py
@@ -3,6 +3,7 @@
 from datalex_core.dbt.emit import emit_dbt, build_sources_yaml, build_models_yaml, EmitReport
 from datalex_core.dbt.manifest import import_manifest, write_import_result, ImportResult
 from datalex_core.dbt.sync import sync_dbt_project, SyncReport, TableSyncRecord
+from datalex_core.dbt.catalog import CatalogIndex, load_catalog, default_catalog_path
 
 __all__ = [
     "emit_dbt",
@@ -15,4 +16,7 @@ __all__ = [
     "sync_dbt_project",
     "SyncReport",
     "TableSyncRecord",
+    "CatalogIndex",
+    "load_catalog",
+    "default_catalog_path",
 ]